requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pyzmq
sqlalchemy
psycopg2-binary
langchain
langchain-community
openai
psutil
backoff

